+++ NEWS.md
@@ -3,6 +3,7 @@
 * Stop letting roof break things
 * New location for pre-computed numbers
 * Re-work field goal model. Takes into account improvements in accuracy (especially on long kicks) and gives probability > 0 on kicks of over 70 yards
+* Update models and internals to  be compatible with newer xgboost (>= v3) versions. (#41)
 
 # nfl4th 1.0.4